ask user if they want to change the polytope type to an H polytope or not
Initially, if the polytope is not an H-polytope but the method is max_ellipsoid, an error is thrown saying that the max_ellipsoid is not applicable for V- or Z- polytopes. This change is made so that if the user entered the wrong polytope type by mistake but actually wants to use the max_ellipsoid method for H-polytope, the user gets an option to change it.